requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
httpx==0.25.2
transformers==4.36.2
torch==2.1.1
onnxruntime==1.16.3
beautifulsoup4==4.12.2
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
